Parametrize export views
