+++ src/lexicon/exports/ExportView.vue
@@ -1,5 +1,12 @@
-<script lang="ts" setup></script>
+<script setup lang="ts">
+import ExportViewTemplate from "@/exports/ExportViewTemplate.vue";
+
+defineProps<{
+  id: string;
+  path: string;
+}>();
+</script>
 
 <template>
-  <div>TODO</div>
+  <ExportViewTemplate type="lexicon" :id :path />
 </template>
