--- src/lexicon/exports/LexiconResult.vue
@@ -1,5 +1,0 @@
-<script lang="ts" setup></script>
-
-<template>
-  <div>TODO</div>
-</template>
